Remove unused parseHttpMessage function from httsserver.c
Add at least some sort of client support for SECURELOCATION.UPNP.ORG to libupnp. Actually all it now does, is that it puts value of that header to discovery struct returned when discovery result is received.

Authorization still wasn't working correctly in gatedevice. Now CONTROL_POINT_HALF_AUTHORIZED is returned if SSL is not used and it is not required to be used. Authorization error is returned right away in actionhandler only if CONTROL_POINT_NOT_AUTHORIZED is returned from AuthorizeControlPoint.

git-svn-id: http://us000253/svn/IGDv2@378 55a18753-f1c3-4722-b33a-e097f9cf6a53