modifications in "gatedevice.c" :
Before this, if SSL was not used, a CP which has a Public Role was able to invoke all actions.
According to the specifications, a CP with a Public Role can do only actions which require Public Role when the SSL connection is not established.
This adds this constraint in the "AuthorizeControlPoint" function and delete the "CONTROL_POINT_HALF_AUTHORIZED" variable.